+++ Sources/Utils/DebugUtils/DebugGui.cs
@@ -2,16 +2,30 @@
 // Author: igor.zavoychinskiy@gmail.com
 // This software is distributed under Public domain license.
 
+using KSPDev.LogUtils;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using UnityEngine;
 
 namespace KSPDev.DebugUtils {
 
 /// <summary>Helper class to deal with the debug GUI functionality.</summary>
 /// <seealso cref="DebugAdjustableAttribute"/>
 public static class DebugGui {
+
+  /// <summary>Game object to use to host the debug dialogs.</summary>
+  /// <seealso cref="MakePartDebugDialog"/>
+  static GameObject dialogsRoot {
+    get {
+      if (_dialogsRoot == null) {
+        _dialogsRoot = new GameObject("dialogsRoot-" + LibraryLoader.assemblyVersionStr);
+      }
+      return _dialogsRoot;
+    }
+  }
+  static GameObject _dialogsRoot;
 
   /// <summary>Metadata about the member that is available for debugging.</summary>
   public struct DebugMemberInfo {
@@ -43,6 +57,34 @@ public static class DebugGui {
           fieldInfo = f
         })
         .ToList();
+  }
+
+  /// <summary>Creates a debug dialog for the parts.</summary>
+  /// <param name="title">The titile of the dialog.</param>
+  /// <param name="dialogWidth">
+  /// The width of the dialog. If omitted, then the code will decide.
+  /// </param>
+  /// <param name="valueColumnWidth">
+  /// The width of the value changing controls. If omitted, then the code will decide.
+  /// </param>
+  /// <returns>The created dialog.</returns>
+  /// <seealso cref="DestroyPartDebugDialog"/>
+  /// <seealso cref="DebugAdjustableAttribute"/>
+  public static PartDebugAdjustmentDialog MakePartDebugDialog(
+      string title, float? dialogWidth = null, float? valueColumnWidth = null) {
+    var dlg = dialogsRoot.AddComponent<PartDebugAdjustmentDialog>();
+    dlg.dialogTitle = title;
+    dlg.dialogWidth = dialogWidth ?? dlg.dialogWidth;
+    dlg.dialogValueSize = valueColumnWidth ?? dlg.dialogValueSize;
+    DebugEx.Info("Created debug dialog: {0}", title);
+    return dlg;
+  }
+
+  /// <summary>Destroys the debug dialog.</summary>
+  /// <param name="dlg">The dialog to destroy.</param>
+  /// <seealso cref="MakePartDebugDialog"/>
+  public static void DestroyPartDebugDialog(PartDebugAdjustmentDialog dlg) {
+    UnityEngine.Object.Destroy(dlg);
   }
 }